Committing in preparation for changing to effayoh defined political entities approach.
Originally I thought it would make sense to define different types for
data source defined political entities. After some more work, I realized
that it would be better to instead define a set of political entities
for the effayoh package. Data sources would define the mapping from their
political entities to the effayoh political entities, and the effayoh
political entities would be used to group or split the political
entites. One advantage of this approach is that it avoids conflicting
groupings or splits between data sources. It also avoids ambiguity in
grouping or splitting if, say, source S groups countries A and B, but
source T neither groups not splits either A or B. This approach also
decouples the data sources. There is no need to map political entities
between sources S and T. As the number of data sources grows this will
be a huge benefit.