Fixed bug in defining the FBSElement constants.
All the elements from SEED to PRODUCTION had a value of "Food" for
element. This caused the BaseFBSMunger to miss the target elements. The
mistake was conspicuous because the production of all countries was 0.

Fixed bug in call to super().__init__:

Part of the super machinery automatically passes the self argument, so
the explicit self passed in the code was redundant and caused an error
as super().__init__ was receiving 3 arguments instead of the expected 2.