Refactored PSDCommodity to subclass tuple rather than str for consistency.
Fixed bug in PSDCommodityGroup __init__ method.

Added attribute_commodity_conversion instance attribute to PSDMunger.

Added PSDMunger.set_attribute_commodity_conversion method.

Fixed bug in PSDMunger.muge method. A dict with dict values was being
passed as an argument to the lambda function. This resulted in a
ValueError when the interpreter tried to add an int and a dict.