Fixed bug in the lambda functions.
When the loop variable is referenced in the lambda function without a
default parameter, the value is not computed until after the loop
finishes execution. At that point the value of the reference is the last
value of the loop variable. Thus, every lambda function had the same
value of factor. The solution is to define the lambda function with a
second parameter with a default value of the loop variable. This value
is then computed at the time the lambda function is created and results
in the intended behavior.