Added logic to modify the production of the epicenter.
Modified logic that computes the initial change in consumption to absorb
up to the parameterized fraction of current consumption rather than of
the residual shock.

Changed model behavior when it attempts to inject a parameter into the
namespace that is already there. Instead of raising an error, it prints
a message.